KEA-83 create a navigation config file in order to type navigation & move AppNavigator to App.tsx
